Extract DFB acquire-release analysis
Move the reserve/push and wait/pop ownership logic out of ttl-insert-cb-sync into a shared DFB acquire-release analysis utility. Reuse the shared analysis from ttl-insert-cb-sync so later PipeNet analyses can consume the same ownership rules without duplicating them.
